+++ example/cars_server/controller.dart
@@ -1,8 +1,7 @@
 import 'dart:async';
 
+import 'package:json_api/server.dart';
 import 'package:json_api_document/json_api_document.dart';
-import 'package:json_api_server/json_api_server.dart';
-import 'package:json_api_server/src/pagination/page.dart';
 import 'package:uuid/uuid.dart';
 
 import 'dao.dart';