+++ lib/json_api.dart
@@ -1,3 +1,3 @@
-export 'package:json_api/src/client.dart';
-export 'package:json_api/src/response.dart';
-export 'package:json_api/src/status_code.dart';
+export 'package:json_api/src/client/client.dart';
+export 'package:json_api/src/client/response.dart';
+export 'package:json_api/src/client/status_code.dart';