+++ lib/src/client/response.dart
@@ -1,5 +1,5 @@
+import 'package:json_api/src/client/status_code.dart';
 import 'package:json_api/src/nullable.dart';
-import 'package:json_api/src/status_code.dart';
 import 'package:json_api_document/json_api_document.dart';
 
 /// A response returned by JSON:API client