+++ lib/src/server/pagination/pagination_strategy.dart
@@ -1,0 +1,5 @@
+import 'package:json_api/src/server/pagination/page.dart';
+
+abstract class PaginationStrategy {
+  Page getPage(Map<String, List<String>> query);
+}