+++ test/functional/delete_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:json_api/json_api.dart';
-import 'package:json_api_server/json_api_server.dart';
+import 'package:json_api/server.dart';
 import 'package:test/test.dart';
 
 import '../../example/cars_server.dart';