+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- `JsonApiClient` has `const` constructor
+### Changed
+- Consolidated the Client, the Server and the Document in one single library
 
 ## [1.0.1] - 2019-04-05
 ### Fixed